+++ the-buddy-system/tests/ui/fail/enum.rs
@@ -1,0 +1,8 @@
+use the_buddy_system::PageFrame;
+
+#[derive(PageFrame)]
+enum NotAStruct {
+    A,
+}
+
+fn main() {}