+++ the-buddy-system/tests/ui/fail/generic.rs
@@ -1,0 +1,6 @@
+use the_buddy_system::PageFrame;
+
+#[derive(PageFrame)]
+struct Generic<T>(T);
+
+fn main() {}