+++ the-buddy-system/tests/ui/fail/unit_struct.rs
@@ -1,0 +1,6 @@
+use the_buddy_system::PageFrame;
+
+#[derive(PageFrame)]
+struct Unit;
+
+fn main() {}